Changes seed file to address a bug that created too many builds.
The Factory that was used to create the seed Assets were creating too many duplicated Build objects. By switching from a Factory to a normal create method, the problem is avoided because it will not attempt to create a new Build, but instead it will get one randomly from the ones already created.

The seeds file also creates 600 Inspections for a randomly chosen set of assets, and 300 Inspection Orders.