Refactors Inspection model
Necessary refactoring to reflect accurate information on a given Asset. Attribute name changed from gas certificate to emissions certificate.

Adds necessary attributes to include the inspector comments on exterior, interior, and engine status.

Includes relationship with Insurance Companies so that an Asset can be insured by a different company at the moment of the inspection.