Adds Comments controller and corresponding haml views.
Adds Comment controller with only three actions to ensure that a comment will always have a parent, and once a comment is created it cannot be destroyed.

Adds partial for one comment, and a partial for creating a comment. Both views are to be used by the parent commentable models.