Adds validations to Model and Build models.
Adds validation to the Model model to ensure that each object is unique within the context of the Make to which it belongs.

Adds validation to the Build model to ensure uniqueness of each build combination. To avoid wrong uniqueness check, the validation has to be scoped so that the same kind can be combined with different bodies.