Adds several changes to inspection model related classes.
Adds a check to the inspection method that updated the inspection count
on the parent asset after the inspection is succesfully saved. This
check was required because in certain cases the inspection was being
created without a parent asset and the method would break.

Adds the inspection related show page.

Adds validations to ensure presence of the parent asset, and to ensure
uniqueness of the inspection number in relation of the asset being
inspected.