Modifies models to support contract query depending on role of entity
The entity can act as a lessee, and it can also act as a cosigner, as
such the models have to support queries for each of the roles the entity
can make.

The cosigner model now validates that a cosigner cannot be the lessee of
the same contract. The entity model adds a method to retrieve the
contracts for which this entity is the lessee.