Adds local uploading feature.
Adds several changes related to the uploading feature. One thing to note
is that the feature works locally but not to the cloud. In the model it
adds the uniqueness validation so that a user cannot upload the same
document more than once.

Adds two partial views, one is the form to upload new files, and the
other one is related to displaying the uploaded documents.