Refactors comments to dry views and ajaxify actions.
Extract a partial view that renders the commenting feature. The new
partial contains both the list of comments and the form to add a new
comment. This refactoring adds consistency in the views and dries the
code base.

Adds support for asynchronous actions thru ajax calls. Adds styling to
the comments section.