Adds styling to the inspections views.
Front end styling with html tags and css to use the responsive grid
provided by the foundation library. The layout depends on the content
but it's subject to the 12 grid system of the framework.

The buttons were styled using glyphicons from the AwesomeFont library to
increase readability and reduce clutter.