Adds styling of different view components, mainly contracts.
Adds styling and layouts for the contract views. Refactors contracts
views by extracting a partial view that renders the entity information.

Forms start making use of the notices partial refactored beforehand. The
main page layout factors the section to display flash notices and errors
so that other pages don't have to, this reduces code and cleans the
views.