Changes the landing pages for authenticated and unathenticated users.
Changes the routes file so that authenticated users get a landing page
withing the app, and the unathenticated user get sent to the landing
page.

It adds two controllers a dashboard controllers that will be in charge
of getting the information to render an information dashboard, and a
pages controller that will be in charge of the static pages.