Fixes issues related to validations.
Fixes #25 #28 #29 #30 #32. This issues had to do with improper
validation of different fields when creating or updating an asset.

Adds a new test to validate that an asset is created with a make that
belongs to the same model as the one assigned to the new asset.

Renames an instance variable on the spec so that it didn't conflict
later on with method names from the framework.