Adds eager loading when searching for an asset or an inspection.
Since so many models are required to render the show view for assets and
inspections, the app has to query the DB several times. This slows the
response to the user and dimishes the overall user experience.

Eager loading was added so that only 1 query gets made requesting all
the information so that response times get reduced.