Fixes bug on migration that changed type of attributes on inspections
On commit 23ca051 a migration changed the type of three attributes and
changed the name of a fourth one from the inspection model. That change
was made to support business requirements. A typo was made and as a
result the migrations didn't ran and the app broke when a user tried to
access any resource related to the inspection model.

This commit erases that migration and adds an updated fixed version.