Adds rake file with task to upload parametric data.
Adds a file with a series of rake tasks to aid in reading parametric
data from csv files and uploading it to the production DB. In order to
upload data correctly, the Make model needs a special field named
runt_code. This code is needed to keep the data consistent with runt's
oficial data.