Modifies inspection_order model and adds proper associations to models.
Changes the inspection_order model to reflect the structure of an
inspection order made by a user. Each inspection order belongs to an
asset, so that each individual inspection can be tracked thru it's
order.

The inspection order makes use of the assm gem, to model the state
behavior of an order and the action it can trigger when entering or
leaving each state.