Changes redirect after contract creation and adds new action.
Previously when a new contract was created, the app would redirect to
the list of contracts, but UX research indicated the the flow should
move the user to the show page of the newly created contract. From
there, the user could add a new contract or a new asset to the new contract.