Adds decorators to separate view logic from view templates
To support such decoration the Draper gem was introduced as a
dependency. The new decorator objects will live in the decorators
folder, with their test counterparts on the spec/decorators folder.

Assets' show view was cleaned as a result of the decorator class, but
the controller had to add the decorator to the asset object retrieved
from the DB.

Changes to the overall structure of the asset's show view were
introduced an a new slider that either takes the asset's documents, or a
placeholder to keep a consistent design.