Improves commenting behaviour and fixes bug.
There was a bug where a brand new comment would not be rendered by the
js because the DOM didn't contain the required div, unless a full page
reload was executed.

The fix checks for the required id and if doesn't exist it creates it,
and then attaches the new comment to the created or found element.