Reverts back to showing an empty table when no record was found.
In a previous release when the user searched for a particular contract
and didn't find it, the app would render a 404 page. This became very
confusing for the user since they believed they'd found a bug o that
something was broken.

Now it displays an empty table and a message that no contracts were
found.