Adds destroy behavior on models with dependent associations.
When destroying a record the has_many side of the relationship would
remain orphan in the db. Since those models contain sensitive
information a condition was added so that a model with a has_many
association cannot be destroy if it still has dependent records
associated with it.

The entity model declares an explicit inverse association with contracts
so that a lessee has many leases_on contracts, and it's related destroy
behavior.