Fixes bug related to deletion of inspections and inspection orders
When an inspection order or an inspection were deleted the associated
model would break if a relationship was established already. To prevent
such issue, checks we're implemented such that before destroying an
inspection, it reverts the state of the inspection order; and the
Inspection order can only be removed if it hasn't been inspected and an
inspection record has been associated with it.